# src/main/java/br/com/robson/robssohex/transportlayers/admin/dto/AdminRoleSummary.java
package br.com.robson.robssohex.transportlayers.admin.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AdminRoleSummary {
    private String id;
    private String name;
    private String description;
}