# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Robsson SSO API
  version: 1.0.0
  description: API de autenticação e controle de acesso baseada em RBAC com JWT
servers:
  - url: http://localhost:8080
paths:
...
  /admin/permissions:
    post:
      tags: [Admin]
      summary: Criar permissão
      description: Cria uma permissão do tipo action/resource usada em `@PreAuthorize`.
      operationId: createPermission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminPermissionRequest" }
            examples:
              exemplo:
                value:
                  resource: "user"
                  action: "read"
      responses:
        "200":
          description: Permissão criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPermissionResponse" }
              examples:
                ok:
                  value:
                    id: "perm_123"
                    resource: "user"
                    action: "read"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [Admin]
      summary: Listar permissões (paginado)
      description: Retorna permissões com paginação e filtro por resource/action.
      operationId: listPermissions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Página de permissões
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPagedResponse" }
              examples:
                ok:
                  value:
                    content:
                      - id: "perm_123"
                        resource: "user"
                        action: "read"
                    page: 0
                    size: 10
                    totalElements: 1
                    totalPages: 1
  /admin/permissions/{id}:
    get:
      tags: [Admin]
      summary: Buscar permissão por id
      description: Retorna uma permissão específica.
      operationId: getPermission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Permissão encontrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPermissionResponse" }
        "400":
          description: Permissão não encontrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    put:
      tags: [Admin]
      summary: Atualizar permissão
      description: Atualiza `resource` e `action` de uma permissão existente.
      operationId: updatePermission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminPermissionRequest" }
            examples:
              exemplo:
                value:
                  resource: "role"
                  action: "write"
      responses:
        "200":
          description: Permissão atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPermissionResponse" }
        "400":
          description: Permissão não encontrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    delete:
      tags: [Admin]
      summary: Remover permissão
      description: Remove uma permissão pelo id.
      operationId: deletePermission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Permissão removida
  /admin/roles:
    post:
      tags: [Admin]
      summary: Criar role
      description: Cria uma role com nome, descrição e permissões opcionais.
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminRoleRequest" }
            examples:
              exemplo:
                value:
                  name: "ADMIN"
                  description: "Acesso total ao painel"
                  permissionIds: ["perm_123", "perm_456"]
      responses:
        "200":
          description: Role criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRoleResponse" }

    get:
      tags: [Admin]
      summary: Listar roles (paginado)
      description: Retorna roles com paginação e filtro por name.
      operationId: listRoles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Página de roles
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPagedResponse" }
              examples:
                ok:
                  value:
                    content:
                      - id: "role_1"
                        name: "ADMIN"
                        description: "Acesso total"
                        permissions: []
                    page: 0
                    size: 10
                    totalElements: 1
                    totalPages: 1
  /admin/me:
    get:
      tags: [Admin]
      summary: Usuário autenticado
      description: Retorna dados do usuário autenticado para montagem de menu.
      operationId: adminMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usuário autenticado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
              examples:
                ok:
                  value:
                    id: "user_1"
                    username: "robson"
                    email: "robson@example.com"
                    ativo: true
                    roles: []
                    permissions: []
  /admin/roles/{id}:
    get:
      tags: [Admin]
      summary: Buscar role por id
      description: Retorna uma role específica com suas permissões.
      operationId: getRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Role encontrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRoleResponse" }

    put:
      tags: [Admin]
      summary: Atualizar role
      description: Atualiza nome, descrição e permissões da role.
      operationId: updateRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminRoleRequest" }
            examples:
              exemplo:
                value:
                  name: "SUPPORT"
                  description: "Suporte com acesso limitado"
                  permissionIds: ["perm_123"]
      responses:
        "200":
          description: Role atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRoleResponse" }

    delete:
      tags: [Admin]
      summary: Remover role
      description: Remove uma role pelo id.
      operationId: deleteRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Role removida
  /admin/roles/{roleId}/permissions/{permissionId}:
    post:
      tags: [Admin]
      summary: Vincular permissão à role
      description: Adiciona uma permissão específica à role.
      operationId: addPermissionToRole
      security:
        - bearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema: { type: string }
        - name: permissionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Role atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRoleResponse" }

    delete:
      tags: [Admin]
      summary: Remover permissão da role
      description: Remove o vínculo entre role e permissão.
      operationId: removePermissionFromRole
      security:
        - bearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema: { type: string }
        - name: permissionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Role atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRoleResponse" }
  /admin/users:
    get:
      tags: [Admin]
      summary: Listar usuários (paginado)
      description: Lista usuários com paginação, filtro por username/email e ativo.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: active
          in: query
          required: false
          schema: { type: boolean }
          description: Filtra por ativos/inativos
      responses:
        "200":
          description: Página de usuários
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPagedResponse" }
              examples:
                ok:
                  value:
                    content:
                      - id: "user_1"
                        username: "robson"
                        email: "robson@example.com"
                        ativo: true
                        roles: []
                        permissions: []
                    page: 0
                    size: 10
                    totalElements: 1
                    totalPages: 1
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Buscar usuário por id
      description: Retorna dados do usuário, roles e permissões diretas.
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
  /admin/users/{id}/activate:
    patch:
      tags: [Admin]
      summary: Ativar usuário
      description: Marca o usuário como ativo.
      operationId: activateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário ativado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
  /admin/users/{id}/deactivate:
    patch:
      tags: [Admin]
      summary: Desativar usuário
      description: Marca o usuário como inativo.
      operationId: deactivateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário desativado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
  /admin/users/{userId}/roles/{roleId}:
    post:
      tags: [Admin]
      summary: Vincular role ao usuário
      description: Adiciona uma role ao usuário.
      operationId: addRoleToUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: roleId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }

    delete:
      tags: [Admin]
      summary: Remover role do usuário
      description: Remove o vínculo entre usuário e role.
      operationId: removeRoleFromUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: roleId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
  /admin/users/{userId}/permissions/{permissionId}:
    post:
      tags: [Admin]
      summary: Vincular permissão direta ao usuário
      description: Adiciona uma permissão direta ao usuário.
      operationId: addPermissionToUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: permissionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }

    delete:
      tags: [Admin]
      summary: Remover permissão direta do usuário
      description: Remove o vínculo direto entre usuário e permissão.
      operationId: removePermissionFromUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: permissionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Parâmetros inválidos."
        details:
          type: object
          additionalProperties: true
...
    AdminPagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    AdminPermissionRequest:
      type: object
      required: [resource, action]
      properties:
        resource:
          type: string
          example: user
        action:
          type: string
          example: read
    AdminPermissionResponse:
      type: object
      properties:
        id:
          type: string
        resource:
          type: string
        action:
          type: string
    AdminRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: ADMIN
        description:
          type: string
          example: Acesso total ao painel
        permissionIds:
          type: array
          items:
            type: string
    AdminRoleResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/AdminPermissionResponse"
    AdminRoleSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        ativo:
          type: boolean
        roles:
          type: array
          items:
            $ref: "#/components/schemas/AdminRoleSummary"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/AdminPermissionResponse"